+++ src/vector.ts
@@ -3,4 +3,8 @@ type Vector2 = {
 	y: number
 }
 
-export { Vector2 }
+const compare_to_vec2 = (x: number, y: number, vec: Vector2): boolean => {
+	return x === vec.x && y === vec.y
+}
+
+export { Vector2, compare_to_vec2 }